chore: Update CSS and form fields in Signup and ProfileSettings pages
The CSS in the TagPeople module has been updated to change the minimum height of the usersContainer. In the FighterUserProfile component, the table headers have been modified to remove the "Opponent" column and add a "Venue" column. Additionally, the Signup and ProfileSettings pages have been updated to include a "Nickname" field in the form.

Note: This commit message follows the conventional format of "chore: <description>".